Add. 서버 라우터 분리
