+++ raqqoon-sidedish/.gitignore
@@ -22,3 +22,7 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+#test
+
+/server/serverTest.js
