+++ raqqoon-sidedish/package.json
@@ -27,7 +27,8 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "server": "nodemon ./server/server.js"
+    "server": "nodemon ./server/server.js",
+    "testServer": "nodemon ./server/serverTest.js"
   },
   "eslintConfig": {
     "extends": [
